+++ pkg/receive/handler.go
@@ -50,7 +50,7 @@ type Handler struct {
 	mtx      sync.RWMutex
 	hashring Hashring
 
-	// Metrics
+	// Metrics.
 	forwardRequestsTotal *prometheus.CounterVec
 }